+++ CleanArchitectureBlazor/CleanArchitectureBlazor/Components/Pages/Dashboard.razor
@@ -1,7 +1,6 @@
 @page "/dashboard"
 @using Domain
-@using CleanArchitectureBlazor.Services
-@inject IEventService EventService
+@inject IEventRepository EventService
 @inject IShiftRepository ShiftService
 @inject IStaffRepository StaffService
 @inject IStaffAssignmentRepository StaffAssignmentService
