+++ CleanArchitectureBlazor/CleanArchitectureBlazor/Components/Pages/Events.razor
@@ -1,7 +1,6 @@
 @page "/events"
 @using Domain
-@using CleanArchitectureBlazor.Services
-@inject IEventService EventService
+@inject IEventRepository EventService
 @inject IShiftRepository ShiftService
 @inject NavigationManager Navigation
 @rendermode InteractiveServer
