+++ CleanArchitectureBlazor/CleanArchitectureBlazor/Components/Pages/ShiftDetails.razor
@@ -1,6 +1,5 @@
 @page "/shifts/details/{ShiftId:int}"
 @using Domain
-@using CleanArchitectureBlazor.Services
 @inject IShiftRepository ShiftService
 @inject IStaffAssignmentRepository StaffAssignmentService
 @inject NavigationManager Navigation
