+++ CleanArchitectureBlazor/CleanArchitectureBlazor/Components/Pages/ShiftEdit.razor
@@ -1,6 +1,5 @@
 @page "/shifts/edit/{ShiftId:int}"
 @using Domain
-@using CleanArchitectureBlazor.Services
 @inject IShiftRepository ShiftService
 @inject IStaffRepository StaffService
 @inject IStaffAssignmentRepository StaffAssignmentService
