+++ CleanArchitectureBlazor/CleanArchitectureBlazor/Components/Pages/Shifts.razor
@@ -1,8 +1,7 @@
 @page "/shifts"
 @using Domain
-@using CleanArchitectureBlazor.Services
 @inject IShiftRepository ShiftService
-@inject IEventService EventService
+@inject IEventRepository EventService
 @inject NavigationManager Navigation
 @rendermode InteractiveServer
