+++ CleanArchitectureBlazor/CleanArchitectureBlazor/Components/Pages/Staff.razor
@@ -1,6 +1,4 @@
 @page "/staff"
-@using CleanArchitectureBlazor.Services
-
 @inject IStaffRepository StaffService
 @inject NavigationManager Navigation
 @rendermode InteractiveServer
