+++ CleanArchitectureBlazor/CleanArchitectureBlazor/Components/Pages/StaffDetails.razor
@@ -1,6 +1,5 @@
 @page "/staff/details/{StaffId:int}"
 @using Domain
-@using CleanArchitectureBlazor.Services
 @inject IStaffRepository StaffService
 @inject IStaffAssignmentRepository StaffAssignmentService
 @inject NavigationManager Navigation
